Update CTP Dev database project

+++ database/DeployDatabase/Parker_CTP_Dev/dbo/Tables/CommonParameters.sql
@@ -19,8 +19,8 @@
     [last_modified_by]                VARCHAR (100)  NULL,
     [last_modified_on]                DATETIME       NULL,
     [summary_timeout]                 INT            NULL,
-    [break_minutes] FLOAT NULL, 
-    [lunch_minutes] FLOAT NULL, 
+    [break_minutes]                   FLOAT (53)     NULL,
+    [lunch_minutes]                   FLOAT (53)     NULL,
     CONSTRAINT [PK_CommonParameters_Parameter_Id] PRIMARY KEY CLUSTERED ([parameter_id] ASC),
     CONSTRAINT [FK_CommonParameters_Site_Id] FOREIGN KEY ([site_id]) REFERENCES [dbo].[Asset] ([asset_id])
 );
